2193번 2차시도 bottom_up 성공(범위 탓)

+++ Java/src/baekjoon/algoStudy/Main.java
@@ -6,8 +6,8 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int[][] dp = new int[n][3];
-        dp[0] = new int[]{1, 0, 1};
+        long[][] dp = new long[n][3];
+        dp[0] = new long[]{1, 0, 1};
 
         for (int i = 1; i < n; i++) {
             dp[i][0] = dp[i-1][1] * 2 + dp[i-1][2];
